backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.27.0
supabase==2.10.0
websockets==13.1
pydantic-settings==2.1.0
pydantic==2.6.4
python-dotenv==1.0.0
beautifulsoup4==4.12.3
fastapi-cache2[inmemory]>=0.2.0
